cnss2: Fix rough delay issue after global reset
Roughly delaying by 10 milliseconds after global reset is unsafe.
It should be determined whether the PCIE_SOC_GLOBAL_RESET_V bit of
the register PCIE_SOC_GLOBAL_RESET has been automatically cleared
after setting it. If this bit is automatically cleared, it means
that the global reset is complete and the delay can be exited.

For this fix, if the PCIE_SOC_GLOBAL_RESET_V bit is still not
cleared after 200 milliseconds of global reset, it will be forced
to delay exit.

Change-Id: I5d82ab3845bad6ee19cb68d4be3922fb33df2e76